+++ services/.aws-sam/build.toml
@@ -1,0 +1,15 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.9e77c257-7ce2-401a-b419-ac55398a5b2c]
+codeuri = "/home/anon/Repos/Connected-airlines-org/services/booking/functions"
+runtime = "provided.al2"
+architecture = "arm64"
+handler = "bootstrap"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["NameTestFunction"]
+
+[function_build_definitions.9e77c257-7ce2-401a-b419-ac55398a5b2c.metadata]
+BuildMethod = "go1.x"
+
+[layer_build_definitions]
